Move extension tests to their own file
They are slow and large and I usually don't actually care to run them
when running ddl tests.
